// src/API.ts
export type Task = {
  __typename: "Task",
  id: string,
  title: string,
  description?: string | null,
  createdAt: string,
  updatedAt: string,
};
// ...
export type OnCreateTaskSubscription = {
  onCreateTask?:  {
    __typename: "Task",
    id: string,
    title: string,
    description?: string | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};
// ...
export type OnDeleteTaskSubscription = {
  onDeleteTask?:  {
    __typename: "Task",
    id: string,
    title: string,
    description?: string | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};